requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.3.0
